pyproject.toml
[tool.poetry]
name = "musicgenerationwebservice"
version = "0.1.0"
description = "Generates music in a few different genres"
authors = ["n4mespace <zaxarchuk.d@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7.7"
quart = "*"
hypercorn = "*"
tox = "*"
flask-login = "*"
pytest = "*"
pytest-asyncio = "*"
quart-openapi = "*"
numpy = "*"
quart-compress2 = "^1.1.0"
gpt-2-simple = "^0.7.1"
tensorflow = "1.15.2"
quart-jwt-extended = "^0.0.2"
quart-cors = "^0.3.0"
typing = "^3.7.4"

[tool.poetry.dev-dependencies]
Sphinx = "^3.0.3"
rinohtype = "^0.4.0"
sphinxcontrib-napoleon = "^0.7"

[tool.poetry.scripts]
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

